+++ resources/assets/js/components/enso/bulma/toastr/Notification.vue
@@ -4,25 +4,29 @@
         :enter-active-class="enterClass"
         :leave-active-class="leaveClass"
         @after-leave="destroy">
-        <article :class="['message toastr animated', type ? `is-${type}` : '']"
+        <div :class="['box notification toastr animated', type ? `is-${type}` : '']"
             v-if="show">
-            <div class="message-header">
-                <span class="icon is-small" v-if="icon">
-                    <fa :icon="icons[type]"></fa>
-                </span>
-                <span class="is-pulled-right">
-                    {{ i18n(displayTitle) }}
-                </span>
-                <button class="delete"
-                    @click="close()"
-                    v-if="closeButton">
-                </button>
-            </div>
-            <div class="message-body"
-                v-if="message">
-                {{ i18n(message) }}
-            </div>
-        </article>
+            <button class="delete"
+                @click="close()"
+                v-if="closeButton">
+            </button>
+             <article class="media">
+                <div class="media-left">
+                    <span class="icon is-large">
+                        <fa :icon="icons[type]" size="2x"></fa>
+                    </span>
+                </div>
+                <div class="media-content">
+                    <div class="content">
+                        <p>
+                            <strong>{{ i18n(displayTitle) }}</strong>
+                            <br>
+                            {{ i18n(message) }}
+                        </p>
+                    </div>
+                </div>
+            </article>
+        </div>
     </transition>
 
 </template>
@@ -33,26 +37,26 @@ import Vue from 'vue';
 
 import fontawesome from '@fortawesome/fontawesome';
 import {
-    faArrowCircleRight, faInfoCircle, faCheckCircle, faExclamationCircle, faTimesCircle,
+    faComment, faInfoCircle, faCheckCircle, faExclamationCircle, faTimesCircle,
 } from '@fortawesome/fontawesome-free-solid/shakable.es';
 
 fontawesome.library.add([
-    faArrowCircleRight, faInfoCircle, faCheckCircle, faExclamationCircle, faTimesCircle,
+    faComment, faInfoCircle, faCheckCircle, faExclamationCircle, faTimesCircle,
 ]);
 
-const types = ['default', 'primary', 'info', 'success', 'warning', 'danger'];
+const types = ['message', 'primary', 'info', 'success', 'warning', 'danger'];
 const positions = ['left', 'right', 'center'];
 const icons = {
-    default: null,
-    primary: 'arrow-circle-right',
-    info: 'info-circle',
-    success: 'check-circle',
-    warning: 'exclamation-circle',
-    danger: 'times-circle',
+    message: faComment,
+    primary: faComment,
+    info: faInfoCircle,
+    success: faCheckCircle,
+    warning: faExclamationCircle,
+    danger: faTimesCircle,
 };
 const titles = {
-    default: 'Message',
-    primary: 'Message',
+    message: 'Message',
+    primary: 'Notification',
     info: 'Info',
     success: 'Success',
     warning: 'Warning',
@@ -60,7 +64,7 @@ const titles = {
 };
 
 export default {
-    name: 'Message',
+    name: 'Notification',
 
     props: {
         type: {
@@ -204,8 +208,9 @@ export default {
             margin-left: calc(50% - 150px);
         }
 
-        .toastr.message {
+        .box.toastr.notification {
             width: 300px;
+            padding: 12px;
             margin-bottom: 6px;
             pointer-events: auto;
             position: relative;
@@ -214,13 +219,6 @@ export default {
             -webkit-box-shadow: 0px 0px 5px 1px rgba(133,133,133,1);
             -moz-box-shadow: 0px 0px 5px 1px rgba(133,133,133,1);
             box-shadow: 0px 0px 5px 1px rgba(133,133,133,1);
-
-            .message-header {
-
-                .icon {
-                    margin-right: 8px;
-                }
-            }
         }
     }